Silence cast warning

+++ vulkan-learning/TriangleBasics/TriangleBasics/TriangleBasicsApp.cpp
@@ -665,7 +665,7 @@ private:
 			int width, height;
 			glfwGetWindowSize(window, &width, &height);
 
-			VkExtent2D actualExtent = { width, height };
+			VkExtent2D actualExtent = { static_cast<uint32_t>(width), static_cast<uint32_t>(height) };
 
 			actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
 			actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
